+++ MSURandomizer/Services/UnknownMsuWindowService.cs
@@ -13,19 +13,28 @@ public class UnknownMsuWindowService(
     IMapper mapper,
     IMsuTypeService msuTypeService,
     IMsuCacheService msuCacheService,
-    IMsuLookupService msuLookupService) : ControlService
+    IMsuLookupService msuLookupService,
+    IMsuHardwareService msuHardwareService) : ControlService
 {
     private UnknownMsuWindowViewModel _model = new();
     private const int MaxPathLength = 50;
+    private bool _isHardwareMode;
     
-    public UnknownMsuWindowViewModel InitilizeModel()
+    public UnknownMsuWindowViewModel InitilizeModel(bool isHardwareMode)
     {
+        _isHardwareMode = isHardwareMode;
+        
         List<string> msuTypes = [""];
         msuTypes.AddRange(msuTypeService.MsuTypes.Select(x => x.DisplayName).OrderBy(x => x));
 
         List<MsuDetailsWindowViewModel> msuModels = [];
+
+        var msuList = isHardwareMode ? msuHardwareService.Msus : msuLookupService.Msus;
+        var msus = msuList.Where(x =>
+            x is { IgnoreUnknown: false, MsuType: null, NumUniqueTracks: >= 15 } &&
+            string.IsNullOrEmpty(x.Settings.MsuTypeName));
         
-        foreach (var msu in msuLookupService.Msus.Where(x => x is { IgnoreUnknown: false, MsuType: null, NumUniqueTracks: >= 15 } && string.IsNullOrEmpty(x.Settings.MsuTypeName) ))
+        foreach (var msu in msus)
         {
             var msuModel = mapper.Map<MsuDetailsWindowViewModel>(msu.Settings);
             msuModel.Msu = msu;
@@ -46,6 +55,7 @@ public class UnknownMsuWindowService(
             {
                 _model.HasBeenModified = true;
             };
+            
             msuModels.Add(msuModel);
         }
 
@@ -62,18 +72,29 @@ public class UnknownMsuWindowService(
         foreach (var msuModel in _model.UnknownMsus.Where(x => x.Msu != null))
         {
             mapper.Map(msuModel, msuModel.Msu!.Settings);
+            msuModel.Msu.MsuType = msuTypeService.GetMsuType(msuModel.MsuTypeName);
             msuModel.Msu.Settings.IsUserUnknownMsu = string.IsNullOrEmpty(msuModel.MsuTypeName);
             userOptionsService.UpdateMsuSettings(msuModel.Msu);
-            msuCacheService.Remove(msuModel.Msu.Path, false);
+            
+            if (_isHardwareMode && !string.IsNullOrEmpty(msuModel.MsuTypeName))
+            {
+                msuHardwareService.RefreshMsu(msuModel.MsuPath);
+            }
+            else if (!_isHardwareMode)
+            {
+                msuCacheService.Remove(msuModel.Msu.Path, false);
+            }
         }
 
         userOptionsService.Save();
-        
-        ITaskService.Run(() =>
+
+        if (!_isHardwareMode)
         {
-            msuLookupService.LookupMsus();
-        });
-        
+            ITaskService.Run(() =>
+            {
+                msuLookupService.LookupMsus();
+            });
+        }
     }
 
     public void SaveIgnore()
